+++ Content/about/index.md
@@ -1,0 +1,15 @@
+---
+title: About
+description: Nice to meet you, welcome to my blog. I am a junior developer who started iOS development in 2021. During this time, I worked with small and large teams on the project.
+---
+
+Nice to meet you, welcome to my blog. I am a junior developer who started iOS development in 2021. During this time, I worked with small and large teams on the project.
+
+I am interested in various architectures and technologies while developing iOS. 
+And I like to challenge myself. Challenging gives me a lot of difficulties, but it gives me more experience and a sense of accomplishment.
+
+I want to keep developing this blog and record my growth process.
+
+Although blog posts may be lacking, please take good care of us!
+
+If you have any questions, please feel free to contact us via email!
