--- Content/debate.md
@@ -1,4 +1,0 @@
----
-title: iOS related discussions or issues.
-description: All the iOS-related issues that are emerging these days.
----
