+++ Content/debate/index.md
@@ -1,0 +1,4 @@
+---
+title: Debate
+description: All the iOS-related issues that are emerging these days.
+---
