+++ Content/posts/index.md
@@ -1,0 +1,1 @@
+# Posts
