+++ Resources/static/scripts/Channel_talk/Channel_talk.js
