+++ Sources/Blog/BlogTheme/BlogHTMLFactory.swift
@@ -1,121 +1,3 @@
-//import Publish
-//import Plot
-//
-//struct BlogHTMLFactory: HTMLFactory {
-//    func makeIndexHTML(for index: Index, context: PublishingContext<Blog>) throws -> HTML {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(
-//                for: context.site,
-//                description: "JihoonAHN Blog records my growth process. I share all my knowledge and experiences with you."
-//            ),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .homePage(for: context.allItems(
-//                        sortedBy: \.date,
-//                        order: .descending)
-//                        .filter { $0.sectionID == .posts },
-//                        on: context.site),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//
-//    func makeSectionHTML(for section: Section<Blog>, context: PublishingContext<Blog>) throws -> HTML {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(for: context.site),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .div(.h1(.text(section.title))),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//
-//    func makeItemHTML(for item: Item<Blog>, context: PublishingContext<Blog>) throws -> HTML {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(
-//                for: context.site,
-//                title: item.title,
-//                description: item.description
-//            ),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .post(for: item, on: context.site),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//
-//    func makePageHTML(for page: Page, context: PublishingContext<Blog>) throws -> HTML {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(
-//                for: context.site,
-//                title: page.title,
-//                description: page.description
-//            ),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .page(for: page, context: context),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//
-//    func makeTagListHTML(for page: TagListPage, context: PublishingContext<Blog>) throws -> HTML? {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(
-//                for: context.site,
-//                title: "JiHoonAHN blog All Tags",
-//                description: "All the tags on JiHoonAHN blog | \(context.allTags.map{ $0.string }.joined(separator: ", "))"
-//            ),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .tagList(for: page.tags.reversed(), on: context.site),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//
-//    func makeTagDetailsHTML(for page: TagDetailsPage, context: PublishingContext<Blog>) throws -> HTML? {
-//        HTML(
-//            .lang(context.site.language),
-//            .head(
-//                for: context.site,
-//                title: "Tags | \(page.tag.string)",
-//                description: "\(page.tag.string) | Found \(context.pages.count) posts"
-//            ),
-//            .body(
-//                .grid(
-//                    .header(for: context.site),
-//                    .tagDetail(
-//                        for: context.items(
-//                            taggedWith: page.tag,
-//                            sortedBy: \.date),
-//                        on: context.site,
-//                        title: "\(page.tag.string.capitalized)"
-//                    ),
-//                    .footer(for: context.site)
-//                )
-//            )
-//        )
-//    }
-//}
-
 import Plot
 import Publish
 import Foundation
@@ -127,16 +9,11 @@ struct BlogHTMLFactory: HTMLFactory {
                        context: Publish.PublishingContext<Blog>) throws -> Plot.HTML {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: index, context: context),
             .body {
                 SiteHeader(context: context)
                 Wrapper {
-                    IndexPage(
-                        pageNumber: 1,
-                        items: context.paginatedItems.first ?? [],
-                        context: context
-                    )
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
+                    Script(.src("/static/scripts/Channel_talk.js"))
                 }
                 SiteFooter(context: context)
             }
@@ -147,13 +24,12 @@ struct BlogHTMLFactory: HTMLFactory {
                          context: Publish.PublishingContext<Blog>) throws -> Plot.HTML {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: section, context: context),
             .body {
                 SiteHeader(context: context)
                 Wrapper {
                     H1(section.title)
-                    DebateList(items: section.items, site: context.site, dateFormatter: .blog)
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
+                    Script(.src("/static/scripts/Channel_talk.js"))
                 }
                 SiteFooter(context: context)
             }
@@ -164,11 +40,12 @@ struct BlogHTMLFactory: HTMLFactory {
                       context: Publish.PublishingContext<Blog>) throws -> Plot.HTML {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: item, context: context),
             .body {
                 SiteHeader(context: context)
                 Wrapper {
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
+                    H1(item.title)
+                    Script(.src("/static/scripts/Channel_talk.js"))
                 }
                 SiteFooter(context: context)
             }
@@ -179,12 +56,10 @@ struct BlogHTMLFactory: HTMLFactory {
                       context: Publish.PublishingContext<Blog>) throws -> Plot.HTML {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: page, context: context),
             .body {
                 SiteHeader(context: context)
-                Wrapper {
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
-                }
+                Wrapper (page.body)
                 SiteFooter(context: context)
             }
         )
@@ -194,11 +69,11 @@ struct BlogHTMLFactory: HTMLFactory {
                          context: Publish.PublishingContext<Blog>) throws -> Plot.HTML? {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: page, context: context),
             .body {
                 SiteHeader(context: context)
                 Wrapper {
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
+                    Script(.src("/static/scripts/Channel_talk.js"))
                 }
                 SiteFooter(context: context)
             }
@@ -209,11 +84,11 @@ struct BlogHTMLFactory: HTMLFactory {
                             context: Publish.PublishingContext<Blog>) throws -> Plot.HTML? {
         HTML(
             .lang(context.site.language),
-            .head(for: context.site),
+            .head(for: page, context: context),
             .body {
                 SiteHeader(context: context)
                 Wrapper {
-                    Script(url: "/js/Channel_talk/Channel_talk.js")
+                    Script(.src("/static/scripts/Channel_talk.js"))
                 }
                 SiteFooter(context: context)
             }
