+++ Sources/Blog/BlogTheme/Components/Footer/SiteFooter.swift
@@ -1,0 +1,36 @@
+import Plot
+import Publish
+
+struct SiteFooter: Component {
+    var context: PublishingContext<Blog>
+    
+    var body: Component {
+        Footer {
+            Div {
+                Div {
+                    Div {
+                        Paragraph {
+                            Text("Copyright © JiHoonAHN")
+                        }
+                    }
+                    Div {
+                        Paragraph {
+                            Text("Made with Swift")
+                        }
+                    }
+                }.class("copyright")
+                Div {
+                    List(context.site.socialMediaLinks) { socialMediaLink in
+                        ListItem {
+                            Link(url: socialMediaLink.url) {
+                                Node.i(
+                                    .class(socialMediaLink.icon)
+                                )
+                            }
+                        }.class("social_link")
+                    }
+                }.class("social")
+            }.class("site-footer__inner")
+        }.class("site-footer")
+    }
+}
