--- Sources/Blog/BlogTheme/Components/Script.swift
@@ -1,15 +1,0 @@
-import Plot
-
-struct Script: Component {
-    // MARK: - Properties
-    var url: URLRepresentable
-    
-    // MARK: - Initalizer
-    init(url: URLRepresentable) {
-        self.url = url
-    }
-    
-    var body: Component {
-        Node<HTML.BodyContext>.script(.src(url))
-    }
-}
