+++ Sources/Blog/BlogTheme/Components/Script/Script.swift
@@ -1,0 +1,15 @@
+import Plot
+
+struct Script: Component {
+    // MARK: - Properties
+    var nodes: Node<HTML.ScriptContext>
+    
+    // MARK: - Initalizer
+    init(_ nodes: Node<HTML.ScriptContext>) {
+        self.nodes = nodes
+    }
+    
+    var body: Component {
+        Node<HTML.BodyContext>.script(nodes)
+    }
+}
