+++ Sources/Blog/BlogTheme/DataFormatter.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 extension DateFormatter {
-    static var blog: DateFormatter = {
+    static var blogTime: DateFormatter = {
         let formatter = DateFormatter()
         formatter.dateStyle = .long
         formatter.dateFormat = "yyyy.MM.dd"
