--- Sources/Blog/BlogTheme/Nodes/Common properties /Grid/Node+Grid.swift
@@ -1,17 +1,0 @@
-import Plot
-import Publish
-
-extension Node where Context == HTML.BodyContext {
-    static func grid(_ nodes: Node...) -> Node {
-        return .div(
-            .class("home-template has-cover"),
-            .div(
-                .class("viewport"),
-                .group(nodes)
-            ),
-            .script(
-                .src("/js/Channel_talk/Channel_talk.js")
-            )
-        )
-    }
-}
