+++ Sources/Blog/BlogTheme/Nodes/Head/GoogleFont/GoogleFont.swift
@@ -1,0 +1,21 @@
+import Plot
+
+extension Node where Context == HTML.HeadContext {
+    static func googleFont() -> Node {
+        return .group(
+            .link(
+                .rel(.preconnect),
+                .href("https://fonts.googleapis.com")
+            ),
+            .link(
+                .rel(.preconnect),
+                .href("https://fonts.gstatic.com"),
+                .crossorigin()
+            ),
+            .link(
+                .href("https://fonts.googleapis.com/css2?family=Open+Sans&display=swap"),
+                .rel(.stylesheet)
+            )
+        )
+    }
+}
