+++ Sources/Blog/BlogTheme/Utility/Index+paginatedPath.swift
@@ -1,8 +1,8 @@
-import Publish
-
-extension Index {
-    func paginatedPath(pageIndex: Int) -> Path {
-        guard pageIndex != 0 else { return path }
-        return path.appendingComponent("\(pageIndex + 1)")
-    }
-}
+//import Publish
+//
+//extension Index {
+//    func paginatedPath(pageIndex: Int) -> Path {
+//        guard pageIndex != 0 else { return path }
+//        return path.appendingComponent("\(pageIndex + 1)")
+//    }
+//}
