--- Sources/Blog/PublishSteps/PaginatedPages/Publising+generatePaginatedPages.swift
@@ -1,73 +1,0 @@
-import Publish
-
-extension PublishingStep where Site == Blog {
-    static func generatePaginatedPages() -> Self {
-        .group([
-            .generatePaginatedIndexPages(),
-//            .generatePaginatedTagPages(),
-//            .generatePaginatedTagListPages()
-        ])
-    }
-
-    private static func generatePaginatedIndexPages() -> Self {
-        .step(named: "Generate paginated index pages") { context in
-            context.paginatedItems.indices.dropFirst().forEach { pageIndex in
-                context.addPage(
-                    Page(
-                        path: context.index.paginatedPath(pageIndex: pageIndex),
-                        content: .init(body: .init(components: {
-                            IndexPage(
-                                pageNumber: pageIndex + 1,
-                                items: context.paginatedItems[pageIndex],
-                                context: context
-                            )
-                        }))
-                    )
-                )
-            }
-        }
-    }
-
-//    private static func generatePaginatedTagPages() -> Self {
-//        .step(named: "Generate paginated tag pages") { context in
-//            context.allTags.forEach { tag in
-//                context.paginatedItems(for: tag).indices.dropFirst().forEach { pageIndex in
-//                    context.addPage(
-//                        Page(
-//                            path: context.site.paginatedPath(for: tag, pageIndex: pageIndex),
-//                            content: .init(body: .init(components: {
-//                                TagsPage(
-//                                    selectedTag: tag,
-//                                    pageNumber: pageIndex + 1,
-//                                    items: context.paginatedItems(for: tag)[pageIndex],
-//                                    context: context
-//                                )
-//                            }))
-//                        )
-//                    )
-//                }
-//            }
-//
-//        }
-//    }
-//
-//    private static func generatePaginatedTagListPages() -> Self {
-//        .step(named: "Generate paginated tag list page") { context in
-//            context.paginatedItems.indices.dropFirst().forEach { pageIndex in
-//                context.addPage(
-//                    Page(
-//                        path: context.site.paginatedTagListPath(pageIndex: pageIndex),
-//                        content: .init(body: .init(components: {
-//                            TagsPage(
-//                                selectedTag: nil,
-//                                pageNumber: pageIndex + 1,
-//                                items: context.paginatedItems[pageIndex],
-//                                context: context
-//                            )
-//                        }))
-//                    )
-//                )
-//            }
-//        }
-//    }
-}
