+++ Sources/Blog/SocialMediaLink.swift
@@ -7,18 +7,25 @@ struct SocialMediaLink {
 }
 
 extension SocialMediaLink {
+    static var twitter: SocialMediaLink {
+        return SocialMediaLink(
+            title: "Twitter",
+            url: "https://twitter.com/jihoon_dev",
+            icon: "fa-brands fa-twitter"
+        )
+    }
     static var github: SocialMediaLink {
         return SocialMediaLink(
             title: "GitHub",
             url: "https://github.com/JiHoonAHN",
-            icon: "fab fa-github-square"
+            icon: "fa-brands fa-github"
         )
     }
     static var email: SocialMediaLink {
         return SocialMediaLink(
             title: "Email",
             url: "mailto:official@jihoon.me",
-            icon: "fas fa-envelope-open-text"
+            icon: "fa-solid fa-envelope"
         )
     }
 }
