+++ Sources/Blog/main.swift
@@ -7,10 +7,8 @@ try Blog().publish(using: [
     .installPlugin(.splash(withClassPrefix: "")),
     .addMarkdownFiles(),
     .copyResources(),
-    .generatePaginatedPages(),
     .generateHTML(withTheme: .blog),
     .generateRSSFeed(including: [.posts]),
-    .move404FileForBlog(),
     .generateSiteMap(),
     .deploy(using: .gitHub("JiHoonAHN/Blog"))
 ])
